[AGENT] merge world continuity into evaluator
